pyproject.toml
[project]
name = "youtube-mcp-clients"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.5",
    "langchain-mcp-adapters>=0.1.13",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
